tests/wpt/tests/html/canvas/offscreen/reset/2d.reset.state.word_spacing.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.reset.state.word_spacing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.reset.state.word_spacing</h1>
<p class="desc">check that the state is reset</p>


<script>
var t = async_test("check that the state is reset");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  const default_value = ctx.wordSpacing;

  ctx.wordSpacing = '12px';
  _assert(ctx.wordSpacing == '12px', "ctx.wordSpacing == '12px'");

  ctx.reset();
  _assert(ctx.wordSpacing == default_value, "ctx.wordSpacing == default_value");
  t.done();

});
</script>
